requirements.txt
# Django and DRF
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Environment variables
python-decouple==3.8

# Image handling
Pillow==10.2.0

# Utilities
django-filter==23.5
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0

# Production server
gunicorn==21.2.0

